requirements.txt
Django==1.11.7
django-tastypie==0.14.0
python-dateutil==2.6.1
python-mimeparse==1.6.0
pytz==2017.3
six==1.11.0
